+++ extensions/empty-screenshots/src/index.ts
@@ -2,22 +2,50 @@ import { readdir, stat, rm } from "fs/promises";
 import { join, resolve } from "path";
 import { homedir } from "os";
 import { showToast, Toast, showHUD, getPreferenceValues } from "@raycast/api";
+import { exec } from "child_process";
 
 interface Preferences {
+  screenshotBehavior: string;
   daysToKeep: string;
   screenshotFolder: string;
+  prefixString: string;
+}
+
+async function moveToTrash(filePath: string) {
+  return new Promise((resolve, reject) => {
+    exec(`osascript -e 'tell application "Finder" to delete POSIX file "${filePath}"'`, (error, stdout, stderr) => {
+      if (error) {
+        console.error("Error moving file to Trash:", stderr);
+        reject(error);
+      } else {
+        console.log("File moved to Trash:", stdout);
+        resolve(stdout);
+      }
+    });
+  });
 }
 
 export default async function main() {
-  const { daysToKeep, screenshotFolder } = getPreferenceValues<Preferences>();
+  const { daysToKeep, screenshotFolder, screenshotBehavior, prefixString } = getPreferenceValues<Preferences>();
   const numberOfDays = parseInt(daysToKeep, 10);
-  const folder = resolve(screenshotFolder.replace("~", homedir));
+  const folder = resolve(screenshotFolder.replace("~", homedir()));
 
   try {
     const files = await readdir(folder);
 
     const screenshots = files
-      .filter((file) => file.endsWith(".png") || file.endsWith(".mov"))
+      .filter((file) => {
+        const endsWithValidExtension = file.endsWith(".png") || file.endsWith(".mov");
+        let startsWithValidPrefix = false;
+
+        if (prefixString) {
+          startsWithValidPrefix = file.startsWith(prefixString);
+        } else {
+          startsWithValidPrefix = file.startsWith("CleanShot") || file.startsWith("Screen");
+        }
+
+        return endsWithValidExtension && startsWithValidPrefix;
+      })
       .map((file) => join(folder, file));
 
     const today = new Date();
@@ -34,7 +62,11 @@ export default async function main() {
       const stats = await stat(screenshot);
 
       if (stats.birthtimeMs <= priorDate.getTime()) {
-        await rm(screenshot);
+        if (screenshotBehavior === "0") {
+          await rm(screenshot);
+        } else {
+          await moveToTrash(screenshot);
+        }
         cleanedCount = cleanedCount + 1;
       }
     }