+++ extensions/empty-screenshots/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Empty Screenshot Folder Changelog
 
+## [Update] - 2024-05-14
+
+- Added preference for the user to choose to delete or move to trash
+- Added preference for the user to choose the screenshot folder
+- Added preference for the user to choose the screenshot custom prefix
+- Added option to not protect recent screenshots
+
 ## [Fix] - 2023-12-01
 
 - Added error handling by providing a warning if Raycast doesn't have permission to access the users screenshot folder.